Update create_mesh.py
    xm = (node_coords[elem_nodes[m][0]][0] + node_coords[elem_nodes[m][1]][0])/2
IndexError: only integers, slices (`:`), ellipsis (`...`), numpy.newaxis (`None`) and integer or boolean arrays are valid indices

+++ create_mesh.py
@@ -24,7 +24,7 @@ def create_line():
     
     for ii in range(0, num_elem):
         elem_nodes[ii] = np.array([ii, ii+1])
-    
+    elem_nodes=elem_nodes.astype(int)
     return {"num_elem": num_elem, "node_coords": node_coords, "elem_nodes": elem_nodes}
     
 def create_reflector():
